+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added property `withoutHeaders` to draw the table without any header row (title + filters)
+
 ## [3.6.0] - 2023-10-26
 
 ### Changed